Switched to using byte instead of int

+++ Arduino/sketches/report_temp/sharedstate.h
@@ -2,6 +2,6 @@
 #define SHAREDSTATE_H
 typedef struct {
 	double currentTemperature;
-	int newTemperature;
+	byte newTemperature;
 } SharedState;
 #endif
